Fix bugs in initial recommendations' collection.

+++ src/main/java/org/moera/node/operations/InitialRecommendationOperations.java
@@ -36,6 +36,10 @@ public class InitialRecommendationOperations {
 
     @Scheduled(fixedDelayString = "P1D")
     public void refresh() {
+        if (!jobs.isReady()) {
+            return;
+        }
+
         try (var ignored = requestCounter.allot()) {
             log.info("Refreshing initial recommendations");
             tx.executeWrite(() -> initialRecommendationRepository.deleteExpired(Util.now()));
